--- a/include/tm_port.h
+++ b/include/tm_port.h
@@ -20,6 +20,7 @@ limitations under the License.
 #define TM_ARCH_RV32P       (4) //T-head E907, etc.
 #define TM_ARCH_RV64V       (5) //T-head C906,C910, etc.
 #define TM_ARCH_CSKYV2      (6) //cskyv2 with dsp core
+#define TM_ARCH_X86_SSE2    (7) //x86 sse2
 
 #define TM_OPT0             (0) //default, least code and buf
 #define TM_OPT1             (1) //opt for speed, need more code and buf