--- a/src/tm_layers.c
+++ b/src/tm_layers.c
@@ -30,6 +30,8 @@ limitations under the License.
     #include "arch_rv64v.h"
 #elif TM_ARCH==TM_ARCH_CSKYV2
     #include "arch_cskyv2.h"
+#elif TM_ARCH==TM_ARCH_X86_SSE2
+    #include "arch_x86_sse2.h"
 #else
     #error "UNSUPPORT ARCH!"
 #endif