run mbnet 1000 times for getting avg time

r9-5950x single thread
fedora 36

---

int8
| optlevel | before(ms) | after(ms) |
| -- | -- | -- |
| 0 | 3.407 | 2.985 |
| 1 | 2.719 | 2.607 |

---

fp32
| optlevel | before(ms) | after(ms) |
| -- | -- | -- |
| 0 | 5.43 | 3.322|
| 1 | 4.889 | 2.575 |