@Zepan 2022-10-17
nihui姐姐，看下 #47 ， sunnycase  优化了下sum后处理，加快了些速度，你可以更新下你的代码看看是否提升更多