@nihui 2022-10-17
> nihui姐姐，看下 #47 ， sunnycase 优化了下sum后处理，加快了些速度，你可以更新下你的代码看看是否提升更多

sunnycase的pr中有关后处理的改动看起来并不能加速呀，是c906的改动在加速